+++ src/com/yanncebron/m68kplugin/lang/psi/M68kTokenGroups.java
@@ -232,6 +232,7 @@ public final class M68kTokenGroups {
     NOLIST,
     NOPAGE,
     ODD,
+    OFFSET,
     OPT,
     ORG,
     PAGE,